refactor: move dev rendering logic inside vite plugin
