+++ packages/astro/src/core/endpoint/dev/index.ts
@@ -1,7 +1,7 @@
 import type { EndpointHandler } from '../../../@types/astro';
-import type { SSROptions } from '../../render/dev';
 import { createRenderContext } from '../../render/index.js';
 import { callEndpoint } from '../index.js';
+import type { SSROptions } from '../../../vite-plugin-astro-server';
 
 export async function call(options: SSROptions) {
 	const { env, preload, middleware } = options;
