+++ packages/astro/src/prerender/routing.ts
@@ -1,6 +1,6 @@
 import type { AstroSettings, ComponentInstance, RouteData } from '../@types/astro';
 import { RedirectComponentInstance, routeIsRedirect } from '../core/redirects/index.js';
-import { preload, type DevelopmentEnvironment } from '../core/render/dev/index.js';
+import { preload, type DevelopmentEnvironment } from '../vite-plugin-astro-server/index.js';
 import { getPrerenderStatus } from './metadata.js';
 
 type GetSortedPreloadedMatchesParams = {
