+++ packages/astro/src/vite-plugin-astro-server/css.ts
@@ -1,7 +1,6 @@
-import type { ModuleLoader } from '../../module-loader/index';
-
-import type { RuntimeMode } from '../../../@types/astro.js';
-import { viteID } from '../../util.js';
+import type { ModuleLoader } from '../core/module-loader/index';
+import type { RuntimeMode } from '../@types/astro.js';
+import { viteID } from '../core/util.js';
 import { isBuildableCSSRequest } from './util.js';
 import { crawlGraph } from './vite.js';
