+++ packages/astro/src/vite-plugin-astro-server/util.ts
