+++ packages/astro/src/vite-plugin-astro-server/vite.ts
@@ -1,9 +1,8 @@
-import type { ModuleLoader, ModuleNode } from '../../module-loader/index';
-
+import type { ModuleLoader, ModuleNode } from '../core/module-loader/index';
 import npath from 'node:path';
-import { SUPPORTED_MARKDOWN_FILE_EXTENSIONS } from '../../constants.js';
-import { unwrapId } from '../../util.js';
 import { isCSSRequest } from './util.js';
+import { unwrapId } from '../core/util.js';
+import { SUPPORTED_MARKDOWN_FILE_EXTENSIONS } from '../core/constants.js';
 
 /**
  * List of file extensions signalling we can (and should) SSR ahead-of-time
